добавлена простая валидация
правки к дз 1 урока, пока без регулярных выражений

+++ Lesson_1_2/Controllers/CardsController.cs
@@ -3,6 +3,8 @@ using Lesson_1_2.DAL.Repositories;
 using Lesson_1_2.DAL.Responses;
 using Lesson_1_2.DAL.DTO;
 using Lesson_1_2.Requests;
+using Lesson_1_2.Validation.Validators;
+using Lesson_1_2.Validation.Service;
 using Microsoft.AspNetCore.Authorization;
 using AutoMapper;
 
@@ -15,11 +17,24 @@ namespace Lesson_1_2.Controllers
     {
         private ICardsRepository Repository;
         private readonly IMapper Mapper;
-
-        public CardsController(ICardsRepository repository, IMapper mapper)
+        private ICreateCardRequestValidator CreateCardRequestValidator;
+        private IUpdateCardRequestValidator UpdateCardRequestValidator;
+        private IDeleteCardRequestValidator DeleteCardRequestValidator;
+        private IGetByIdCardRequestValidator GetByIdCardRequestValidator;
+        public CardsController(
+            ICardsRepository repository, 
+            IMapper mapper,
+            ICreateCardRequestValidator createCardRequestValidator,
+            IUpdateCardRequestValidator updateCardRequestValidator,
+            IDeleteCardRequestValidator deleteCardRequestValidator,
+            IGetByIdCardRequestValidator getByIdCardRequestValidator)
         {
             Mapper = mapper;
             Repository = repository;
+            CreateCardRequestValidator = createCardRequestValidator;
+            UpdateCardRequestValidator = updateCardRequestValidator;
+            DeleteCardRequestValidator = deleteCardRequestValidator;
+            GetByIdCardRequestValidator = getByIdCardRequestValidator;
         }
 
         [HttpGet("get/all")]
@@ -44,6 +59,13 @@ namespace Lesson_1_2.Controllers
         public IActionResult GetById([FromRoute] int id)
         {
             var request = new GetByIdCardRequest(id);
+            var validation = new OperationResult<GetByIdCardRequest>(GetByIdCardRequestValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             var card = Repository.GetById(request);
 
             var response = new GetAllCardsResponse()
@@ -60,6 +82,12 @@ namespace Lesson_1_2.Controllers
         public IActionResult Create([FromRoute] long number, [FromRoute] string name, [FromRoute] DateTimeOffset date, [FromRoute] string type)
         {
             var request = new CreateCardRequest(number, name, date, type);
+            var validation = new OperationResult<CreateCardRequest>(CreateCardRequestValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
 
             Repository.Create(request);
 
@@ -70,6 +98,12 @@ namespace Lesson_1_2.Controllers
         public IActionResult Update([FromRoute] int id, [FromRoute] long number, [FromRoute] string name, [FromRoute] DateTimeOffset date, [FromRoute] string type)
         {
             var request = new UpdateCardRequest(id, number, name, date, type);
+            var validation = new OperationResult<UpdateCardRequest>(UpdateCardRequestValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
 
             Repository.Update(request);
 
@@ -80,6 +114,12 @@ namespace Lesson_1_2.Controllers
         public IActionResult Delete([FromRoute] int id)
         {
             var request = new DeleteCardRequest(id);
+            var validation = new OperationResult<DeleteCardRequest>(DeleteCardRequestValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
 
             Repository.Delete(request);
